Added machinery to compile custom models
