+++ earthsim/models/custom_tools.py
@@ -4,6 +4,8 @@ from bokeh.core.properties import Instance, List
 from bokeh.core.enums import Dimensions
 from bokeh.models import Tool, ColumnDataSource
 
+from . import _CUSTOM_MODELS
+
 fpath = os.path.dirname(__file__)
 
 
@@ -13,7 +15,7 @@ class CheckpointTool(Tool):
     the RestoreTool to restore the data to a previous state.
     """
 
-    __implementation__ = os.path.join(fpath, 'custom_tools.ts')
+    __implementation__ = os.path.join(fpath, 'checkpoint_tool.ts')
 
     sources = List(Instance(ColumnDataSource))
 
@@ -24,7 +26,7 @@ class RestoreTool(Tool):
     checkpoint created by the CheckpointTool
     """
 
-    __implementation__ = os.path.join(fpath, 'custom_tools.ts')
+    __implementation__ = os.path.join(fpath, 'restore_tool.ts')
 
     sources = List(Instance(ColumnDataSource))
 
@@ -34,6 +36,11 @@ class ClearTool(Tool):
     Clears the data on the supplied ColumnDataSources.
     """
 
-    __implementation__ = os.path.join(fpath, 'custom_tools.ts')
+    __implementation__ = os.path.join(fpath, 'clear_tool.ts')
 
     sources = List(Instance(ColumnDataSource))
+
+
+_CUSTOM_MODELS['earthsim.models.custom_tools.CheckPointTool'] = CheckpointTool
+_CUSTOM_MODELS['earthsim.models.custom_tools.RestoreTool'] = RestoreTool
+_CUSTOM_MODELS['earthsim.models.custom_tools.ClearTool'] = ClearTool
